change aligment to common property
