+++ sycl/doc/extensions/experimental/sycl_ext_oneapi_annotated_ptr.asciidoc
@@ -613,44 +613,6 @@ increment and decrement operator of annotated_ref. The annotations are applied
 when the object `T` is loaded and stored to the memory.
 |===
 
-=== Properties
-
-Below is a list of compile-time constant properties supported with
-`annotated_ptr`. Other extensions can define additional compile-time constant
-properties that can be supported with `annotated_ptr`. Runtime properties
-are not supported.
-
-```c++
-namespace sycl::ext::oneapi::experimental {
-struct alignment_key {
-  template<int K>
-  using value_t = property_value<alignment_key, std::integral_constant<int, K>>;
-};
-
-template<int K>
-inline constexpr alignment_key::value_t<K> alignment;
-
-template<>
-struct is_property_key<alignment_key> : std::true_type {};
-
-template<typename T, typename PropertyListT>
-struct is_property_key_of<
-  alignment_key, annotated_ptr<T, PropertyListT>> : std::true_type {};
-template<typename T, typename PropertyListT>
-} // namespace sycl::ext::oneapi::experimental
-```
---
-[options="header"]
-|====
-| Property | Description
-|`alignment`
-|
-This property is an assertion by the application that the `annotated_ptr` has
-the given alignment, specified in bytes. The behavior is undefined if the
-pointer value does not have the indicated alignment.
-|====
---
-
 == Issues related to `annotated_ptr`
 
 1) [RESOLVED] Should we allow implicit conversion to base class by default?
