+++ sycl/include/sycl/ext/oneapi/experimental/common_annotated_properties/properties.hpp
@@ -24,10 +24,6 @@ template <typename T, typename PropertyListT> class annotated_arg;
 template <typename T, typename PropertyListT> class annotated_ptr;
 
 //===----------------------------------------------------------------------===//
-//        Common properties of annotated_arg/annotated_ptr
-//===----------------------------------------------------------------------===//
-
-//===----------------------------------------------------------------------===//
 //   Utility type trait for annotated_arg/annotated_ptr deduction guide
 //===----------------------------------------------------------------------===//
 template <typename T, typename PropertyValueT>
@@ -57,6 +53,46 @@ struct check_property_list<T, Prop, Props...>
   static_assert(is_valid_property<T, Prop>::value,
                 "Property is invalid for the given type.");
 };
+
+//===----------------------------------------------------------------------===//
+//        Common properties of annotated_arg/annotated_ptr
+//===----------------------------------------------------------------------===//
+struct alignment_key {
+  template <int K>
+  using value_t = property_value<alignment_key, std::integral_constant<int, K>>;
+};
+
+template <int K> inline constexpr alignment_key::value_t<K> alignment;
+
+template <> struct is_property_key<alignment_key> : std::true_type {};
+
+template <typename T, int W>
+struct is_valid_property<T, alignment_key::value_t<W>>
+    : std::bool_constant<std::is_pointer<T>::value> {};
+
+template <typename T, typename PropertyListT>
+struct is_property_key_of<alignment_key, annotated_ptr<T, PropertyListT>>
+    : std::true_type {};
+
+template <typename T, typename PropertyListT>
+struct is_property_key_of<alignment_key, annotated_arg<T, PropertyListT>>
+    : std::true_type {};
+
+namespace detail {
+
+template <> struct PropertyToKind<alignment_key> {
+  static constexpr PropKind Kind = PropKind::Alignment;
+};
+
+template <> struct IsCompileTimeProperty<alignment_key> : std::true_type {};
+
+template <int N> struct PropertyMetaInfo<alignment_key::value_t<N>> {
+  static constexpr const char *name = "sycl-alignment";
+  static constexpr int value = N;
+};
+
+} // namespace detail
+
 
 } // namespace experimental
 } // namespace oneapi
